refactor: extract useMagicCircle hook to split file under 10k chars
- Move Canvas logic (draw, touch, timer, scoring) to src/hooks/useMagicCircle.ts (7.3KB)
- MagicCircleCanvas.tsx reduced to UI-only (5KB)
- All files under 10,000 character limit
- Hook provides: canvasRef, states, handlers, startPoint, getRankColor
